add receiving message from peer

+++ incs/connection.hpp
@@ -5,6 +5,6 @@
 
 int createConnection(const std::string& ip, const long long port);
 void sendData(const int sockfd, const std::string& msg);
-const std::string recieveData(int sockfd, long long size);
+const std::string recieveData(int sockfd, int size);
 
-#endif // CONNECTION_HPP
+#endif  // CONNECTION_HPP
